pyproject.toml
[tool.poetry]
name = "repl-nix-botpythondiscord"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "repl_nix_botpythondiscord"}]

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.4.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
